Rename cf-manager file
